+++ project/src/main/java/com/shop/service/ProductService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProductService {
     public List<Product> getAll();
     public Boolean create(Product product);
-    public Product find(Integer id);
+    public Product find(Long id);
     public Boolean update(Product product);
-    public Boolean delete(Integer id);
+    public Boolean delete(Long id);
 }
